Add 白痴 (1) and (2)
This is a particularly difficult word. Its dating back to the 10th century, multiple readings (たわけ、ばか、こけ etc), its shift in semantic meaning from a general meaning "idiot" to the scientific term for "mental retardation" all make it difficult to pinpoint exactly the origins of each sense of the term. My suspicions are that the term originates from Chinese but was given an ateji reading for the Japanese term, with the shift to the on'yomi pronunciation はくち signalling a new scientific meaning of "retardation". Since the pronunciation はくち seems a relatively new development, we might consider the second sense being given the reading "たわけ" The earliest evidence I can find of it being given the pronunciation はくち are in scientific contexts e.g. "白痴教育” (1894), i.e. the first sense. The reference I gave for this second sense however does not give a pronunciation but is used in a scientific sense featuring in a text entitled "心理学" as a translation of the works of Alexander Bain.