Update supplement.csv
Ext source: 六名家詩鈔 6巻 真下穆, 植村義[植村蘆洲]選, 167@4@1

Source is again in Classical Chinese (with kundoku readings). 
Huang's definition, however, appears more restricted than the Japanese sense which is not necessarily tied to school. The general sense of a squad (which of course may be applied to a school in Japanese too) seems to be shared with Classical Chinese.