Add Japanese etymology for 傳統(2)
Having looked through Shen's article and corroborating my own results, I will have to agree with Shen of the earliest appearance as being in Taiyo in 1895. This is also the earliest result given in CHJ.

Looking in Japanese-English/French/Russian/German dictionaries compiled in the 1860s and 1870s, not one lists 伝統 as a lemma. So I'm inclined to believe that Taiyo, being the popular literary journal of its day, may have given the word, originating from the Classics, new life.